refactor: new directory structure
