+++ src/contacts/AccountNameHeader.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import './App.css';
-import Icon from './Icon';
-import * as Lib from './lib';
+import Icon from '../ui-shared/Icon';
+import * as Lib from '../lib';
 
 interface AccountNameHeaderProps {
     account: Lib.Account;
