+++ src/sign-in/SignInHelp.tsx
@@ -1,5 +1,3 @@
-import './App.css';
-
 function SignInHelp() {
     return (
         <div className="row d-flex justify-content-center row-space text-start">
